Fix mim dir
