--- siatune/codebase/mim/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) SI-Analytics. All rights reserved.
-from .task import MIM
-
-__all__ = ['MIM']
