+++ siatune/utils/__init__.py
@@ -5,8 +5,9 @@ from .container import ImmutableContainer
 from .dist import set_env_vars
 from .logger import get_root_logger
 from .mim import get_train_script
+from .setup_env import register_all_modules
 
 __all__ = [
     'ImmutableContainer', 'dump_cfg', 'get_root_logger', 'ref_raw_args',
-    'get_train_script', 'set_env_vars'
+    'get_train_script', 'set_env_vars', 'register_all_modules'
 ]
